services/diarization-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pyannote.audio==3.1.1
python-multipart==0.0.6
torch>=2.0.0
torchaudio>=2.0.0
